+++ libnetwork/drivers/bridge/link_test.go
@@ -11,9 +11,9 @@ import (
 
 func getPorts() []types.TransportPort {
 	return []types.TransportPort{
-		{Proto: types.TCP, Port: uint16(5000)},
-		{Proto: types.UDP, Port: uint16(400)},
-		{Proto: types.TCP, Port: uint16(600)},
+		{Proto: types.TCP, Port: 5000},
+		{Proto: types.UDP, Port: 400},
+		{Proto: types.TCP, Port: 600},
 	}
 }
